requirements.txt
langchain==0.3.27
langchain_community==0.4.1
faiss-cpu==1.12.0
langchain-huggingface==1.0.0

gradio
#sentence-transformers
#unstructured[local-inference]
#python-docx
#pypdf
#openai